+++ src/Spago/Build.hs
@@ -15,6 +15,7 @@ import           Spago.Env
 
 import qualified Crypto.Hash          as Hash
 import qualified Data.List.NonEmpty   as NonEmpty
+import qualified Data.Map             as Map
 import qualified Data.Set             as Set
 import qualified Data.Text            as Text
 import           System.Directory     (getCurrentDirectory)
@@ -27,7 +28,6 @@ import qualified Turtle
 import qualified System.Process       as Process
 import qualified Web.Browser          as Browser
 
-import qualified Spago.Build.Parser   as Parse
 import qualified Spago.Command.Path   as Path
 import qualified Spago.RunEnv         as Run
 import qualified Spago.Config         as Config
@@ -49,19 +49,80 @@ prepareBundleDefaults maybeModuleName maybeTargetPath = (moduleName, targetPath)
     targetPath = fromMaybe (TargetPath "index.js") maybeTargetPath
 
 --   eventually running some other action after the build
-build
-  :: HasBuildEnv env
-  => BuildOptions -> Maybe (RIO Env ())
-  -> RIO env ()
-build BuildOptions{..} maybePostBuild = do
+build :: HasBuildEnv env => Maybe (RIO Env ()) -> RIO env ()
+build maybePostBuild = do
   logDebug "Running `spago build`"
+  BuildOptions{..} <- view (the @BuildOptions)
   Config{..} <- view (the @Config)
   deps <- Packages.getProjectDeps
-  case noInstall of
-    DoInstall -> Fetch.fetchPackages deps
-    NoInstall -> pure ()
-  let allPsGlobs = Packages.getGlobs   deps depsOnly configSourcePaths <> sourcePaths
+  let partitionedGlobs@(Packages.Globs{..}) = Packages.getGlobs deps depsOnly configSourcePaths
+      allPsGlobs = Packages.getGlobsSourcePaths partitionedGlobs <> sourcePaths
       allJsGlobs = Packages.getJsGlobs deps depsOnly configSourcePaths <> sourcePaths
+
+      checkImports = do
+        maybeGraph <- view (the @Graph)
+        for_ maybeGraph $ \(Purs.ModuleGraph moduleGraph) -> do
+              let
+                matchesGlob :: Sys.FilePath -> SourcePath -> Bool
+                matchesGlob path sourcePath =
+                  Glob.match (Glob.compile (Text.unpack (unSourcePath sourcePath))) path
+
+                isProjectFile :: Sys.FilePath -> Bool
+                isProjectFile path =
+                  any (matchesGlob path) (fromMaybe [] projectGlobs)
+
+                projectModules :: [ModuleName]
+                projectModules =
+                  map fst
+                    $ filter (\(_, Purs.ModuleGraphNode{..}) -> isProjectFile (Text.unpack graphNodePath))
+                    $ Map.toList moduleGraph
+
+                getImports :: ModuleName -> Set ModuleName
+                getImports = maybe Set.empty (Set.fromList . graphNodeDepends) . flip Map.lookup moduleGraph
+
+                -- All package modules that are imported from our project files
+                importedPackageModules :: Set ModuleName
+                importedPackageModules =
+                  Set.difference
+                    (foldMap getImports projectModules)
+                    (Set.fromList projectModules)
+
+                getPackageFromPath :: Text -> Maybe PackageName
+                getPackageFromPath path =
+                  fmap fst
+                    $ find (\(_, sourcePath) -> matchesGlob (Text.unpack path) sourcePath)
+                    $ Map.toList depsGlobs
+
+                defaultPackages :: Set PackageName
+                defaultPackages = Set.singleton (PackageName "psci-support")
+
+                importedPackages :: Set PackageName
+                importedPackages =
+                  Set.fromList
+                    $ mapMaybe (getPackageFromPath . graphNodePath <=< flip Map.lookup moduleGraph)
+                    $ Set.toList importedPackageModules
+
+                dependencyPackages :: Set PackageName
+                dependencyPackages = Set.fromList dependencies
+
+              let
+                unusedPackages =
+                  fmap packageName
+                    $ Set.toList
+                    $ Set.difference dependencyPackages
+                    $ Set.union defaultPackages importedPackages
+
+                transitivePackages =
+                  fmap packageName
+                    $ Set.toList
+                    $ Set.difference importedPackages dependencyPackages
+
+              unless (null unusedPackages) $ do
+                logWarn $ display $ Messages.unusedDependency unusedPackages
+
+              unless (null transitivePackages) $ do
+                die [ display $ Messages.sourceImportsTransitiveDependency transitivePackages ]
+
 
       buildBackend globs = do
         case alternateBackend of
@@ -82,6 +143,7 @@ build BuildOptions{..} maybePostBuild = do
               shell backendCmd empty >>= \case
                 ExitSuccess   -> pure ()
                 ExitFailure n -> die [ "Backend " <> displayShow backend <> " exited with error:" <> repr n ]
+        checkImports
 
       buildAction globs = do
         env <- Run.getEnv
@@ -155,7 +217,8 @@ repl newPackages sourcePaths pursArgs depsOnly = do
 
         let dependencies = [ PackageName "effect", PackageName "console", PackageName "psci-support" ] <> newPackages
 
-        config <- Config.makeTempConfig dependencies Nothing [] Nothing
+        config <- Run.withPursEnv NoPsa $ do
+          Config.makeTempConfig dependencies Nothing [] Nothing
 
         Run.withInstallEnv' (Just config) (replAction purs)
   where
@@ -168,45 +231,38 @@ repl newPackages sourcePaths pursArgs depsOnly = do
           [ "The package called 'psci-support' needs to be installed for the repl to work properly."
           , "Run `spago install psci-support` to add it to your dependencies."
           ]
-      let globs = Packages.getGlobs deps depsOnly (configSourcePaths <> sourcePaths)
+      let
+        globs =
+          Packages.getGlobsSourcePaths $ Packages.getGlobs deps depsOnly (configSourcePaths <> sourcePaths)
       Fetch.fetchPackages deps
       runRIO purs $ Purs.repl globs pursArgs
 
 
 -- | Test the project: compile and run "Test.Main"
 --   (or the provided module name) with node
-test
-  :: HasBuildEnv env
-  => Maybe ModuleName -> BuildOptions -> [BackendArg]
-  -> RIO env ()
-test maybeModuleName buildOpts extraArgs = do
+test :: HasBuildEnv env => Maybe ModuleName -> [BackendArg] -> RIO env ()
+test maybeModuleName extraArgs = do
   let moduleName = fromMaybe (ModuleName "Test.Main") maybeModuleName
-  Config.Config { alternateBackend, configSourcePaths } <- view (the @Config)
-  liftIO (foldMapM (Glob.glob . Text.unpack . unSourcePath) configSourcePaths) >>= \paths -> do
-    results <- forM paths $ \path -> do
-      content <- readFileBinary path
-      pure $ Parse.checkModuleNameMatches (encodeUtf8 $ unModuleName moduleName) content
-    if or results
-      then do
+  Config.Config { alternateBackend } <- view (the @Config)
+  maybeGraph <- view (the @Graph)
+  -- We check if the test module is included in the build and spit out a nice error if it isn't (see #383)
+  for_ maybeGraph $ \(ModuleGraph moduleMap) -> case Map.lookup moduleName moduleMap of
+    Nothing -> die [ "Module '" <> (display . unModuleName) moduleName <> "' not found! Are you including it in your build?" ]
+    Just _ -> do
         sourceDir <- Turtle.pwd
         let dirs = RunDirectories sourceDir sourceDir
-        runBackend alternateBackend dirs moduleName (Just "Tests succeeded.") "Tests failed: " buildOpts extraArgs
-      else do
-        die [ "Module '" <> (display . unModuleName) moduleName <> "' not found! Are you including it in your build?" ]
+        runBackend alternateBackend dirs moduleName (Just "Tests succeeded.") "Tests failed: " extraArgs
 
 
 -- | Run the project: compile and run "Main"
 --   (or the provided module name) with node
-run
-  :: HasBuildEnv env
-  => Maybe ModuleName -> BuildOptions -> [BackendArg]
-  -> RIO env ()
-run maybeModuleName buildOpts extraArgs = do
+run :: HasBuildEnv env => Maybe ModuleName -> [BackendArg] -> RIO env ()
+run maybeModuleName extraArgs = do
   Config.Config { alternateBackend } <- view (the @Config)
   let moduleName = fromMaybe (ModuleName "Main") maybeModuleName
   sourceDir <- Turtle.pwd
   let dirs = RunDirectories sourceDir sourceDir
-  runBackend alternateBackend dirs moduleName Nothing "Running failed; " buildOpts extraArgs
+  runBackend alternateBackend dirs moduleName Nothing "Running failed; " extraArgs
 
 
 -- | Run the select module as a script: init, compile, and run the provided module
@@ -217,7 +273,7 @@ script
   -> [PackageName]
   -> ScriptBuildOptions
   -> RIO env ()
-script modulePath tag packageDeps opts@ScriptBuildOptions{..} = do
+script modulePath tag packageDeps opts = do
   logDebug "Running `spago script`"
   absoluteModulePath <- fmap Text.pack (makeAbsolute (Text.unpack modulePath))
   currentDir <- Turtle.pwd
@@ -240,26 +296,16 @@ script modulePath tag packageDeps opts@ScriptBuildOptions{..} = do
 
   let dependencies = [ PackageName "effect", PackageName "console", PackageName "prelude" ] <> packageDeps
 
-  config <- Config.makeTempConfig dependencies Nothing [ SourcePath absoluteModulePath ] tag
+  config <- Run.withPursEnv NoPsa $ do
+    Config.makeTempConfig dependencies Nothing [ SourcePath absoluteModulePath ] tag
 
   let runDirs :: RunDirectories
       runDirs = RunDirectories scriptDirPath currentDir
 
-  Run.withBuildEnv' (Just config) NoPsa (runAction runDirs)
+  Run.withBuildEnv' (Just config) NoPsa buildOpts (runAction runDirs)
   where
-    runAction dirs = do
-      let
-        buildOpts = BuildOptions
-          { shouldClear = NoClear
-          , shouldWatch = BuildOnce
-          , allowIgnored = DoAllowIgnored
-          , sourcePaths = []
-          , withSourceMap = WithoutSrcMap
-          , noInstall = DoInstall
-          , depsOnly = AllSources
-          , ..
-          }
-      runBackend Nothing dirs (ModuleName "Main") Nothing "Script failed to run; " buildOpts []
+    buildOpts = fromScriptOptions defaultBuildOptions opts
+    runAction dirs = runBackend Nothing dirs (ModuleName "Main") Nothing "Script failed to run; " []
 
 
 data RunDirectories = RunDirectories { sourceDir :: FilePath, executeDir :: FilePath }
@@ -273,13 +319,13 @@ runBackend
   -> ModuleName
   -> Maybe Text
   -> Text
-  -> BuildOptions
   -> [BackendArg]
   -> RIO env ()
-runBackend maybeBackend RunDirectories{ sourceDir, executeDir } moduleName maybeSuccessMessage failureMessage buildOpts@BuildOptions{pursArgs} extraArgs = do
+runBackend maybeBackend RunDirectories{ sourceDir, executeDir } moduleName maybeSuccessMessage failureMessage extraArgs = do
   logDebug $ display $ "Running with backend: " <> fromMaybe "nodejs" maybeBackend
+  BuildOptions{ pursArgs } <- view (the @BuildOptions)
   let postBuild = maybe (nodeAction $ Path.getOutputPath pursArgs) backendAction maybeBackend
-  build buildOpts (Just postBuild)
+  build (Just postBuild)
   where
     fromFilePath = Text.pack . Turtle.encodeString
     runJsSource = fromFilePath (sourceDir Turtle.</> ".spago/run.js")
@@ -329,7 +375,7 @@ bundleApp withMain maybeModuleName maybeTargetPath noBuild buildOpts usePsa =
   let (moduleName, targetPath) = prepareBundleDefaults maybeModuleName maybeTargetPath
       bundleAction = Purs.bundle withMain (withSourceMap buildOpts) moduleName targetPath
   in case noBuild of
-    DoBuild -> Run.withBuildEnv usePsa $ build buildOpts (Just bundleAction)
+    DoBuild -> Run.withBuildEnv usePsa buildOpts $ build (Just bundleAction)
     NoBuild -> Run.getEnv >>= (flip runRIO) bundleAction
 
 -- | Bundle into a CommonJS module
@@ -356,25 +402,24 @@ bundleModule maybeModuleName maybeTargetPath noBuild buildOpts usePsa = do
             Right _ -> logInfo $ display $ "Make module succeeded and output file to " <> unTargetPath targetPath
             Left (n :: SomeException) -> die [ "Make module failed: " <> repr n ]
   case noBuild of
-    DoBuild -> Run.withBuildEnv usePsa $ build buildOpts (Just bundleAction)
+    DoBuild -> Run.withBuildEnv usePsa buildOpts $ build (Just bundleAction)
     NoBuild -> Run.getEnv >>= (flip runRIO) bundleAction
 
 
 -- | Generate docs for the `sourcePaths` and run `purescript-docs-search build-index` to patch them.
 docs
-  :: (HasLogFunc env, HasConfig env)
+  :: HasBuildEnv env
   => Maybe Purs.DocsFormat
-  -> [SourcePath]
-  -> Packages.DepsOnly
   -> NoSearch
   -> OpenDocs
   -> RIO env ()
-docs format sourcePaths depsOnly noSearch open = do
+docs format noSearch open = do
   logDebug "Running `spago docs`"
+  BuildOptions { sourcePaths, depsOnly } <- view (the @BuildOptions)
   Config{..} <- view (the @Config)
   deps <- Packages.getProjectDeps
   logInfo "Generating documentation for the project. This might take a while..."
-  Purs.docs docsFormat $ Packages.getGlobs deps depsOnly configSourcePaths <> sourcePaths
+  Purs.docs docsFormat $ Packages.getGlobsSourcePaths (Packages.getGlobs deps depsOnly configSourcePaths) <> sourcePaths
 
   when isHTMLFormat $ do
     when (noSearch == AddSearch) $ do
@@ -406,16 +451,14 @@ docs format sourcePaths depsOnly noSearch open = do
     openLink link = liftIO $ Browser.openBrowser (Text.unpack link)
 
 -- | Start a search REPL.
-search
-  :: (HasPurs env, HasLogFunc env, HasConfig env)
-  => RIO env ()
+search :: HasBuildEnv env => RIO env ()
 search = do
   Config{..} <- view (the @Config)
   deps <- Packages.getProjectDeps
 
   logInfo "Building module metadata..."
 
-  Purs.compile (Packages.getGlobs deps Packages.AllSources configSourcePaths)
+  Purs.compile (Packages.getGlobsSourcePaths (Packages.getGlobs deps Packages.AllSources configSourcePaths))
     [ PursArg "--codegen"
     , PursArg "docs"
     ]