+++ src/Spago/Purs.hs
@@ -1,5 +1,6 @@
 module Spago.Purs
   ( compile
+  , graph
   , repl
   , bundle
   , docs
@@ -7,11 +8,14 @@ module Spago.Purs
   , parseDocsFormat
   , findFlag
   , DocsFormat(..)
+  , ModuleGraph(..)
+  , ModuleGraphNode(..)
   ) where
 
 import           Spago.Prelude
 import           Spago.Env
 
+import qualified Data.ByteString.Lazy as BSL
 import qualified Data.Text      as Text
 import qualified Data.Text.Encoding as Text.Encoding
 import qualified Data.Text.Encoding.Error as Text.Encoding
@@ -26,19 +30,43 @@ compile
   => [SourcePath] -> [PursArg]
   -> RIO env ()
 compile sourcePaths extraArgs = do
-  PursCmd purs <- view (the @PursCmd)
-  logDebug $ "Compiling with " <> displayShow purs
+  PursCmd { purs, psa } <- view (the @PursCmd)
+  logDebug $ "Compiling with " <> displayShow (fromMaybe purs psa)
   let
     paths = Text.intercalate " " $ surroundQuote <$> map unSourcePath sourcePaths
     args  = Text.intercalate " " $ map unPursArg extraArgs
-    cmd = purs <> " compile " <> args <> " " <> paths
+    pursCompile = purs <> " compile"
+    cmd = fromMaybe pursCompile psa <> " " <> args <> " " <> paths
   runWithOutput cmd
     "Build succeeded."
     "Failed to build."
 
+graph
+  :: (HasPurs env, HasLogFunc env)
+  => [SourcePath]
+  -> RIO env (Either Text ModuleGraph)
+graph sourcePaths = do
+  PursCmd { purs } <- view (the @PursCmd)
+  logDebug $ "Getting module graph with " <> displayShow purs
+  let
+    paths = Text.intercalate " " $ surroundQuote <$> map unSourcePath sourcePaths
+    cmd = purs <> " graph " <> paths
+  Turtle.Bytes.shellStrictWithErr cmd empty >>= \case
+    (ExitSuccess, out, _err) -> do
+      let graphText = Text.Encoding.decodeUtf8With lenientDecode out
+          parsed = decode $ BSL.fromStrict $ encodeUtf8 graphText
+
+      pure $ case parsed of
+        Nothing -> Left $ Messages.failedToParseCommandOutput cmd graphText
+        Just p -> Right p
+
+    (_, _out, err) ->
+      pure $ Left $ "Failed to run `" <> cmd <> "`. Error was:\n" <> tshow err
+
+
 repl :: HasPurs env => [SourcePath] -> [PursArg] -> RIO env ()
 repl sourcePaths extraArgs = do
-  PursCmd purs <- view (the @PursCmd)
+  PursCmd { purs } <- view (the @PursCmd)
   let paths = Text.intercalate " " $ surroundQuote <$> map unSourcePath sourcePaths
       args = Text.intercalate " " $ map unPursArg extraArgs
       cmd = purs <> " repl " <> paths <> " " <> args