+++ src/Spago/Types.hs
@@ -76,6 +76,7 @@ data Force = Force | NoForce
 data IncludeTransitive = IncludeTransitive | NoIncludeTransitive
 
 newtype ModuleName = ModuleName { unModuleName :: Text }
+  deriving newtype (Eq, FromJSON, FromJSONKey, Ord)
 newtype TargetPath = TargetPath { unTargetPath :: Text }
 newtype SourcePath = SourcePath { unSourcePath :: Text }
   deriving newtype (Show, Dhall.FromDhall)
@@ -122,6 +123,7 @@ data NoBuild = NoBuild | DoBuild
 
 -- | Flag to skip the automatic installation of libraries on build
 data NoInstall = NoInstall | DoInstall
+  deriving Eq
 
 -- Should we clear the screen on rebuild?
 data ClearScreen = DoClear | NoClear
@@ -145,6 +147,29 @@ data BuildOptions = BuildOptions
   , beforeCommands :: [Text]
   , thenCommands   :: [Text]
   , elseCommands   :: [Text]
+  }
+
+defaultBuildOptions :: BuildOptions
+defaultBuildOptions = BuildOptions
+  { shouldClear = NoClear
+  , shouldWatch = BuildOnce
+  , allowIgnored = DoAllowIgnored
+  , sourcePaths = []
+  , withSourceMap = WithoutSrcMap
+  , noInstall = DoInstall
+  , depsOnly = AllSources
+  , pursArgs = []
+  , beforeCommands = []
+  , thenCommands = []
+  , elseCommands = []
+  }
+
+fromScriptOptions :: BuildOptions -> ScriptBuildOptions -> BuildOptions
+fromScriptOptions opts ScriptBuildOptions{..} = opts
+  { pursArgs = pursArgs
+  , beforeCommands = beforeCommands
+  , thenCommands = thenCommands
+  , elseCommands = elseCommands
   }
 
 -- TODO: Figure out how `Watch` would work for `spago script` and include it
@@ -173,10 +198,31 @@ data PublishConfig = PublishConfig
   , publishRepository :: Text
   } deriving (Show, Generic)
 
+data PursCmd = PursCmd
+  { purs :: Text
+  , psa :: Maybe Text
+  , compilerVersion :: Version.SemVer
+  } deriving (Generic)
+
 newtype Jobs = Jobs Int
 newtype ConfigPath = ConfigPath Text
-newtype PursCmd = PursCmd Text
 newtype GitCmd = GitCmd Text
 newtype BowerCmd = BowerCmd Text
 
 data GlobalCache = GlobalCache !GHC.IO.FilePath !(Maybe CacheFlag)
+
+newtype ModuleGraph = ModuleGraph { unModuleGraph :: Map ModuleName ModuleGraphNode }
+  deriving newtype (FromJSON)
+
+data ModuleGraphNode = ModuleGraphNode
+  { graphNodePath :: Text
+  , graphNodeDepends :: [ModuleName]
+  } deriving (Generic)
+
+instance FromJSON ModuleGraphNode where
+  parseJSON = withObject "ModuleGraphNode" $ \o ->
+    ModuleGraphNode
+      <$> o .: "path"
+      <*> o .: "depends"
+
+type Graph = Maybe ModuleGraph